refactor: Move chat logic to its own class
